Switched to original MPU6050 library and refactored for readability

+++ src/utils/utils.cpp
@@ -5,8 +5,8 @@ void blink(int blinkRate, int blinkDuration) {
   pinMode(LED_BUILTIN, OUTPUT);
   for (int i = 0; i < blinkDuration; i += blinkRate) {
     digitalWrite(LED_BUILTIN, HIGH);
-    delay((float) blinkRate / 2);
+    delay((double) blinkRate / 2);
     digitalWrite(LED_BUILTIN, LOW);
-    delay((float) blinkRate / 2);
+    delay((double) blinkRate / 2);
   }
 }
